refactor: Реорганизация
