+++ src/components/app-header/app-header.tsx
@@ -1,16 +1,14 @@
+import { useMemo } from "react";
 import { NavLink, useLocation } from "react-router-dom";
 import { useMediaQuery } from "react-responsive";
 import Hamburger from "hamburger-react";
 
-import {
-  Logo,
-  BurgerIcon,
-  ListIcon,
-  ProfileIcon,
-} from "@ya.praktikum/react-developer-burger-ui-components";
+import { Logo } from "@ya.praktikum/react-developer-burger-ui-components";
+import { LogoSmall } from "../../assets/images/logo-small";
+import { Sidebar } from "./components/sidebar/sidebar";
+import { NavLinkWithIcon } from "./components/nav-link-with-icon/nav-link-with-icon";
 import { useAppSelector } from "../../services/hooks";
 import { userAuthorized } from "../../utils/utils";
-import { LogoSmall } from "../../assets/images/logo-small";
 import { urls } from "../../utils/urls";
 
 import styles from "./app-header.module.css";
@@ -19,75 +17,70 @@ interface IAppHeader {
   showSidebar: boolean;
   toggleSidebar: () => void;
 }
-
-const setActive = ({ isActive }: { isActive: Boolean }) =>
-  "text text_type_main-default ml-2 " +
-  (isActive ? styles["menu__item_active"] : styles["menu__item"]);
-
-const buttonType = (link: string, active: string) => {
-  if (active === link) return "primary";
-  return "secondary";
-};
-
 export function AppHeader({ showSidebar, toggleSidebar }: IAppHeader) {
   const location = useLocation();
   const { user } = useAppSelector((state) => state.authUser);
+  const authorized = useMemo(() => userAuthorized(user), [user]);
 
   const isMobileOrTablet = useMediaQuery({ query: "(max-width: 1280px)" });
 
   return (
-    <header className={styles["header"]}>
-      <nav className={styles["header-inner"]}>
-        <ul className={`${styles["menu-left"]} mb-4 mt-4`}>
-          <li className={`${styles["menu__item"]} ml-5 mr-5 mb-5 mt-5`}>
-            <BurgerIcon type={buttonType(urls.home, location.pathname)} />
-            <NavLink
-              className={setActive}
-              to={urls.home}
-              state={{ from: location.pathname }}
-            >
-              Конструктор
+    <>
+      <header className={styles["header"]}>
+        <nav className={styles["header-inner"]}>
+          <ul className={`${styles["menu-left"]} mb-4 mt-4`}>
+            <li className={`${styles["menu__item"]} ml-5 mr-5 mb-5 mt-5`}>
+              <NavLinkWithIcon
+                url={urls.home}
+                pathname={location.pathname}
+                text="Конструктор"
+                iconType="home"
+              />
+            </li>
+            <li className={`${styles["menu__item"]} ml-5 mr-5 mb-5 mt-5`}>
+              <NavLinkWithIcon
+                url={urls.feed}
+                pathname={location.pathname}
+                text="Лента заказов"
+                iconType="feed"
+              />
+            </li>
+          </ul>
+          <div className={styles["logo"]}>
+            <NavLink to={urls.home} state={{ from: location }}>
+              {isMobileOrTablet ? <LogoSmall /> : <Logo />}
             </NavLink>
-          </li>
-          <li className={`${styles["menu__item"]} ml-5 mr-5 mb-5 mt-5`}>
-            <ListIcon type={buttonType(urls.feed, location.pathname)} />
-            <NavLink
-              className={setActive}
-              to={urls.feed}
-              state={{ from: location.pathname }}
-              end
-            >
-              Лента заказов
-            </NavLink>
-          </li>
-        </ul>
-        <div className={styles["logo"]}>
-          <NavLink to={urls.home} state={{ from: location }}>
-            {isMobileOrTablet ? <LogoSmall /> : <Logo />}
-          </NavLink>
-        </div>
-        <div className={styles["hamburger"]}>
-          <Hamburger
-            rounded
-            direction="right"
-            onToggle={toggleSidebar}
-            toggled={showSidebar}
-          />
-        </div>
-        <ul className={`${styles["menu-right"]} mb-4 mt-4`}>
-          <li className={`${styles["menu__item"]} ml-5 mr-5 mb-5 mt-5`}>
-            <ProfileIcon type={buttonType(urls.profile, location.pathname)} />
-            <NavLink
-              className={setActive}
-              to={urls.profile}
-              state={{ from: location.pathname }}
-              end
-            >
-              {userAuthorized(user) ? <>Личный кабинет</> : <>Войти</>}
-            </NavLink>
-          </li>
-        </ul>
-      </nav>
-    </header>
+          </div>
+          <div className={styles["hamburger"]}>
+            <Hamburger
+              rounded
+              direction="right"
+              onToggle={toggleSidebar}
+              toggled={showSidebar}
+            />
+          </div>
+          <ul className={`${styles["menu-right"]} mb-4 mt-4`}>
+            <li className={`${styles["menu__item"]} ml-5 mr-5 mb-5 mt-5`}>
+              {authorized ? (
+                <NavLinkWithIcon
+                  url={urls.profile}
+                  pathname={location.pathname}
+                  text="Личный кабинет"
+                  iconType="profile"
+                />
+              ) : (
+                <NavLinkWithIcon
+                  url={urls.login}
+                  pathname={location.pathname}
+                  text="Войти"
+                  iconType="profile"
+                />
+              )}
+            </li>
+          </ul>
+        </nav>
+      </header>
+      <Sidebar showSidebar={showSidebar} toggleSidebar={toggleSidebar} />
+    </>
   );
 }
